requirements.txt
gradio
transformers
torch
accelerate
bitsandbytes

pillow
pytesseract
PyPDF2
numpy